+++ src/irdl/__init__.py
@@ -7,6 +7,6 @@ from .ista import MiracleDataset as MiracleDataset
 from .ista import SrirachaDataset as SrirachaDataset
 from .sofa import BrasRs8Dataset as BrasRs8Dataset
 from .sofa import FabianDataset as FabianDataset
-from .sofacoustics import HutubsDataset as HutubsDataset
+from .sofa import HutubsDataset as HutubsDataset
 
 __all__ = [dataset_class.__name__ for dataset_class in _get_dataset_classes(_sys.modules[__name__])]
